+++ src/main/java/com/company/pgi/repository/ICompanyRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.company.pgi.model.Company;
 
-public interface ICompanyRepository extends JpaRepository<Company, Integer> {
+public interface ICompanyRepository extends JpaRepository<Company, Long> {
 
 }
