+++ src/main/java/com/company/pgi/repository/person/IPersonRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.company.pgi.model.Person;
 
-public interface IPersonRepository extends JpaRepository<Person, Integer> {
+public interface IPersonRepository extends JpaRepository<Person, Long> {
 
 }
